requirements.txt
pandas==2.2.3
python-dotenv==1.0.0